refactor(tui): extract pi interactive adapter
